+++ tests/scar_add_metadata_toolbox_tests/test_app.py
@@ -1,9 +1,7 @@
-from http import HTTPStatus
 from unittest.mock import patch
 
 import pytest
 from flask import Flask
-from flask.testing import FlaskClient
 
 from scar_add_metadata_toolbox import create_app
 from scar_add_metadata_toolbox.config import (
@@ -59,22 +57,3 @@ def test_cli_version(app_runner):
     result = app_runner.invoke(args=["version"])
     assert result.exit_code == 0
     assert "scar-add-metadata-toolbox version: N/A" in result.output
-
-
-@pytest.mark.usefixtures("app_client")
-def test_health(app_client: FlaskClient):
-    expected_response = {
-        "description": "Server side endpoints for the SCAR Antarctic Digital Database " "(ADD) Metadata Toolbox.",
-        "links": {
-            "about": "https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox",
-            "describedBy": "https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/blob/vN/A/README.md",
-            "self": "http://localhost/meta/health/v1",
-        },
-        "releaseId": "N/A",
-        "status": "pass",
-        "version": 1,
-    }
-    response = app_client.get("/meta/health/v1")
-    assert response.status_code == HTTPStatus.OK
-    assert response.mimetype == "application/json"
-    assert response.json == expected_response